test: speed up unit suite and reduce output
