+++ packages/web/src/components/NotebookPage/NotebookPage.test.tsx
@@ -1,331 +1,15 @@
-import type { ReactNode } from 'react';
-import { Suspense } from 'react';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { act, render, screen, waitFor, within } from '@testing-library/react';
+import { describe, expect, it } from 'vitest';
+import { act, screen, waitFor, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { MemoryRouter, Route, Routes } from 'react-router-dom';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { NotebookPage } from './NotebookPage';
-import { ThemeProvider } from '@/context/ThemeContext';
 import { sessionKeys } from '@/api/queryKeys';
-import type { ProjectDetail, Session } from '@/types';
-
-const PID = 'proj-x';
-const NID = 'nb-1';
-
-function ok(data: unknown) {
-	return new Response(JSON.stringify({ success: true, data }), {
-		headers: { 'content-type': 'application/json' },
-	});
-}
-
-function changeRequestBody(init?: RequestInit): { target_proposal_id?: string } {
-	return JSON.parse(String(init?.body ?? '{}')) as { target_proposal_id?: string };
-}
-
-function runningSession(overrides: Partial<Session> = {}): Session {
-	return {
-		session_id: 'sess-1',
-		notebook_id: NID,
-		project_id: PID,
-		user_id: 'me',
-		status: 'running',
-		started_at: '2025-03-05T14:00:00Z',
-		last_heartbeat: '2025-03-05T14:00:00Z',
-		sandbox_url: 'https://sandbox.example/kernel',
-		can: { attach: true, stop: true },
-		...overrides,
-	} as Session;
-}
-
-interface FetchOptions {
-	role: NonNullable<ProjectDetail['your_role']>;
-	viewerMode?: 'static' | 'applications' | 'ephemeral-sandbox';
-	/** Body of GET .../html; null = 404 (no snapshot captured yet). */
-	html?: string | null;
-	session?: Session;
-	/** Items served by GET /projects/{pid}/sessions (the app page's edit-activity poll). */
-	projectSessions?: Session[];
-	/** The notebook's current head version (staleness comparisons). */
-	headVersion?: string;
-	/** The notebook's source type; `git` models a GitHub-synced notebook. */
-	sourceType?: 'local' | 'git';
-	/** When set, POST .../sessions fails with this error code at this status. */
-	createError?: { code: string; message: string; status: number };
-	sessionResponses?: Session[];
-	computeProfiles?: { name: string; cpu?: number; memory_bytes?: number }[];
-	computeProfile?: string;
-	computeProfileOverride?: 'none' | 'editors';
-	editorSharing?: 'shared' | 'exclusive';
-	editorOwner?: { id: string; activity: 'active' | 'idle' | 'unknown' | 'starting' };
-	editorCanTakeOver?: boolean;
-	editorTransfer?: 'requested' | 'draining' | 'ready';
-	editorStateFailures?: number;
-	editorStateFailOn?: number[];
-	meFailures?: number;
-	sourceControlProviders?: string[];
-	omitSourceControlCapability?: boolean;
-	changeRequestFailures?: number;
-	changeRequestFailOn?: number[];
-	changeRequestFailure?: { code: string; message: string; status: number };
-	mePromise?: Promise<{
-		id: string;
-		email: string;
-		logout_url: null;
-		is_super_admin: boolean;
-	}>;
-}
-
-/** Route every request the page makes to a canned response; return the fetch spy. */
-function makeFetch(opts: FetchOptions) {
-	let takeoverComplete = false;
-	let sessionPostCount = 0;
-	let editorStateRequestCount = 0;
-	let meRequestCount = 0;
-	let changeRequestCount = 0;
-	let successfulChangeRequestCount = 0;
-	let nextChangeRequestNumber = 17;
-	const publications = new Map<
-		string,
-		{ provider: string; number: number; url: string; head_branch: string; head_commit: string }
-	>();
-	const impl = vi.fn(async (input: RequestInfo | URL, init?: RequestInit) => {
-		const url = String(input);
-		const method = init?.method ?? 'GET';
-		if (method === 'POST' && url.endsWith(`/notebooks/${NID}/editor-session/takeover`)) {
-			takeoverComplete = true;
-			return ok(undefined);
-		}
-
-		if (method === 'POST' && url.endsWith(`/notebooks/${NID}/sessions`)) {
-			if (opts.createError) {
-				const { code, message, status } = opts.createError;
-				return new Response(JSON.stringify({ success: false, error: { code, message } }), {
-					status,
-					headers: { 'content-type': 'application/json' },
-				});
-			}
-			const response =
-				opts.sessionResponses?.[sessionPostCount] ?? opts.session ?? runningSession();
-			sessionPostCount += 1;
-			return ok({ ...response, reused: false });
-		}
-		if (method === 'POST' && url.endsWith('/change-requests')) {
-			changeRequestCount += 1;
-			if (
-				changeRequestCount <= (opts.changeRequestFailures ?? 0) ||
-				opts.changeRequestFailOn?.includes(changeRequestCount)
-			) {
-				const failure = opts.changeRequestFailure ?? {
-					code: 'SERVICE_UNAVAILABLE',
-					message: 'GitHub is unavailable',
-					status: 503,
-				};
-				return new Response(
-					JSON.stringify({
-						success: false,
-						error: { code: failure.code, message: failure.message },
-					}),
-					{ status: failure.status, headers: { 'content-type': 'application/json' } },
-				);
-			}
-			successfulChangeRequestCount += 1;
-			const proposalId = `prop-${successfulChangeRequestCount}234567890abcdef`;
-			const body = changeRequestBody(init);
-			const target = body.target_proposal_id
-				? publications.get(body.target_proposal_id)
-				: undefined;
-			const number = target?.number ?? nextChangeRequestNumber++;
-			const changeRequest = {
-				provider: 'github',
-				number,
-				url: `https://github.com/org/repo/pull/${number}`,
-				head_branch: target?.head_branch ?? `marimohub/nb-1/${proposalId}`,
-				head_commit: `head-${successfulChangeRequestCount}`,
-			};
-			publications.set(proposalId, changeRequest);
-			return ok({ proposal_id: proposalId, change_request: changeRequest });
-		}
-		if (url.endsWith(`/sessions/${(opts.session ?? runningSession()).session_id}`)) {
-			return ok(opts.session ?? runningSession());
-		}
-		if (url.endsWith(`/projects/${PID}/sessions`)) {
-			return ok({ items: opts.projectSessions ?? [], next_cursor: null });
-		}
-		if (url.endsWith(`/notebooks/${NID}/html`)) {
-			if (opts.html == null) {
-				return new Response(
-					JSON.stringify({ success: false, error: { code: 'NO_HTML_SNAPSHOT', message: 'none' } }),
-					{ status: 404, headers: { 'content-type': 'application/json' } },
-				);
-			}
-			return new Response(opts.html, {
-				headers: {
-					'content-type': 'text/html',
-					'X-Marimohub-Captured-At': '2025-03-05T14:00:00Z',
-				},
-			});
-		}
-		if (url.endsWith(`/notebooks/${NID}`)) {
-			return ok({
-				meta: {
-					id: NID,
-					title: 'Forecast',
-					author: 'me',
-					...(opts.computeProfile ? { compute_profile: opts.computeProfile } : {}),
-				},
-				source:
-					opts.sourceType === 'git'
-						? {
-								type: 'git',
-								provider: 'github',
-								repo: 'org/repo',
-								branch: 'main',
-								root_path: '',
-								entry_notebook: 'app.py',
-								commit: 'deadbeefcafe0123',
-								last_synced_at: '2026-07-01T10:00:00Z',
-								current_version_id: opts.headVersion ?? 'ver-head',
-							}
-						: { type: 'local', current_version_id: opts.headVersion ?? 'ver-head' },
-			});
-		}
-		if (url.includes('/capabilities')) {
-			const viewerMode = opts.viewerMode ?? 'static';
-			return ok({
-				federation: { available: false },
-				...(opts.omitSourceControlCapability
-					? {}
-					: {
-							source_control: {
-								change_request_providers: opts.sourceControlProviders ?? [],
-							},
-						}),
-				viewer_mode: viewerMode,
-				viewer_session_modes:
-					viewerMode === 'ephemeral-sandbox'
-						? ['app', 'edit']
-						: viewerMode === 'applications'
-							? ['app']
-							: [],
-				limits: {},
-				compute_profiles: opts.computeProfiles ?? [],
-				compute_profile_override: opts.computeProfileOverride ?? 'none',
-				editor_sandbox_sharing: opts.editorSharing ?? 'shared',
-			});
-		}
-		if (url.endsWith('/me')) {
-			meRequestCount += 1;
-			if (meRequestCount <= (opts.meFailures ?? 0)) {
-				return new Response(
-					JSON.stringify({
-						success: false,
-						error: { code: 'SERVICE_UNAVAILABLE', message: 'identity unavailable' },
-					}),
-					{ status: 503, headers: { 'content-type': 'application/json' } },
-				);
-			}
-			return ok(
-				opts.mePromise
-					? await opts.mePromise
-					: { id: 'me', email: 'me@example.com', logout_url: null, is_super_admin: false },
-			);
-		}
-		if (url.endsWith(`/notebooks/${NID}/editor-session`)) {
-			editorStateRequestCount += 1;
-			if (
-				editorStateRequestCount <= (opts.editorStateFailures ?? 0) ||
-				opts.editorStateFailOn?.includes(editorStateRequestCount)
-			) {
-				return new Response(
-					JSON.stringify({
-						success: false,
-						error: { code: 'SERVICE_UNAVAILABLE', message: 'ownership unavailable' },
-					}),
-					{ status: 503, headers: { 'content-type': 'application/json' } },
-				);
-			}
-			const owner = takeoverComplete ? undefined : opts.editorOwner;
-			return ok({
-				sharing: opts.editorSharing ?? 'shared',
-				holder: owner
-					? {
-							session_id: 'sess-owner',
-							user_id: owner.id,
-							status: 'running',
-							started_at: '2025-03-05T14:00:00Z',
-							activity: { state: owner.activity },
-						}
-					: null,
-				can_take_over: opts.editorCanTakeOver ?? !!owner,
-				...(opts.editorTransfer
-					? { transfer: { status: opts.editorTransfer } }
-					: takeoverComplete
-						? { transfer: { status: 'ready' } }
-						: {}),
-			});
-		}
-		if (url.includes('/users')) return ok({});
-		if (url.endsWith(`/projects/${PID}`)) {
-			return ok({ id: PID, name: 'P', description: '', your_role: opts.role });
-		}
-		throw new Error(`unexpected fetch: ${method} ${url}`);
-	});
-	vi.stubGlobal('fetch', impl);
-	return impl;
-}
-
-function renderPage(variant: 'edit' | 'app' = 'edit') {
-	const client = new QueryClient({ defaultOptions: { queries: { retry: false } } });
-	const path =
-		variant === 'app'
-			? `/projects/${PID}/notebooks/${NID}/app`
-			: `/projects/${PID}/notebooks/${NID}`;
-	const wrapper = ({ children }: { children: ReactNode }) => (
-		<MemoryRouter initialEntries={[path]}>
-			<QueryClientProvider client={client}>
-				<ThemeProvider>
-					<Suspense fallback={<div>loading</div>}>{children}</Suspense>
-				</ThemeProvider>
-			</QueryClientProvider>
-		</MemoryRouter>
-	);
-	const view = render(
-		<Routes>
-			<Route path="/projects/:pid/notebooks/:nid" element={<NotebookPage />} />
-			<Route path="/projects/:pid/notebooks/:nid/app" element={<NotebookPage variant="app" />} />
-		</Routes>,
-		{ wrapper },
-	);
-	return { ...view, client };
-}
-
-const sessionPosts = (impl: ReturnType<typeof makeFetch>) =>
-	impl.mock.calls.filter(
-		([url, init]) => String(url).endsWith('/sessions') && init?.method === 'POST',
-	);
-
-beforeEach(() => {
-	// The theme baked into the kernel iframe URL reads from localStorage; clear it
-	// so each test resolves the default (light) unless it opts into dark.
-	localStorage.clear();
-	// jsdom has no matchMedia; Tooltip's mobile check needs it.
-	vi.stubGlobal('matchMedia', (query: string) => ({
-		matches: false,
-		media: query,
-		onchange: null,
-		addEventListener: () => {},
-		removeEventListener: () => {},
-		addListener: () => {},
-		removeListener: () => {},
-		dispatchEvent: () => false,
-	}));
-});
-
-afterEach(() => {
-	vi.unstubAllGlobals();
-	vi.restoreAllMocks();
-});
+import {
+	NID,
+	PID,
+	makeFetch,
+	renderPage,
+	runningSession,
+	sessionPosts,
+} from './NotebookPage.testWorld';
 
 describe('NotebookPage viewer modes', () => {
 	it('starts shared editing without requesting exclusive ownership state', async () => {
@@ -655,769 +339,5 @@ describe('NotebookPage viewer modes', () => {
 			container.querySelector('iframe[src="https://sandbox.example/kernel?theme=light"]'),
 		).not.toBeNull();
 		expect(sessionPosts(impl)).toHaveLength(1);
-	});
-});
-
-describe('NotebookPage git-synced editor', () => {
-	const gitEditSession = (overrides: Partial<Session> = {}) =>
-		runningSession({ source_version_id: 'ver-head', ...overrides });
-
-	it('shows the updated-on-GitHub banner when the session trails the head', async () => {
-		makeFetch({
-			role: 'editor',
-			sourceType: 'git',
-			session: gitEditSession({ source_version_id: 'ver-old' }),
-			headVersion: 'ver-head',
-		});
-		renderPage();
-
-		await waitFor(() =>
-			expect(screen.getByText(/updated in its git repository/)).toBeInTheDocument(),
-		);
-		expect(screen.getByText('Restart to update')).toBeInTheDocument();
-	});
-
-	it('shows no banner when the session serves the synced head', async () => {
-		makeFetch({ role: 'editor', sourceType: 'git', session: gitEditSession() });
-		const { container } = renderPage();
-
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-		expect(screen.queryByText(/updated in its git repository/)).toBeNull();
-	});
-
-	it('header shows the repo chip whose popover links to the source on GitHub', async () => {
-		const user = userEvent.setup();
-		makeFetch({ role: 'editor', sourceType: 'git', session: gitEditSession() });
-		renderPage();
-
-		await user.click(
-			await screen.findByRole('button', { name: 'Synced from a git repository — details' }),
-		);
-		const popover = await screen.findByRole('dialog');
-		expect(within(popover).getByRole('link', { name: 'org/repo' })).toHaveAttribute(
-			'href',
-			'https://github.com/org/repo',
-		);
-		expect(within(popover).getByRole('link', { name: 'deadbee' })).toHaveAttribute(
-			'href',
-			'https://github.com/org/repo/commit/deadbeefcafe0123',
-		);
-		expect(within(popover).getByRole('link', { name: /View source on GitHub/ })).toHaveAttribute(
-			'href',
-			'https://github.com/org/repo/blob/deadbeefcafe0123/app.py',
-		);
-	});
-
-	it('the app view shows no repo chip', async () => {
-		makeFetch({
-			role: 'editor',
-			sourceType: 'git',
-			session: gitEditSession({ mode: 'app' }),
-		});
-		const { container } = renderPage('app');
-
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-		expect(
-			screen.queryByRole('button', { name: 'Synced from a git repository — details' }),
-		).toBeNull();
-	});
-
-	it('shows the banner without a restart CTA when the caller cannot stop the session', async () => {
-		makeFetch({
-			role: 'editor',
-			sourceType: 'git',
-			session: gitEditSession({
-				source_version_id: 'ver-old',
-				can: { attach: true, stop: false },
-			}),
-			headVersion: 'ver-head',
-		});
-		renderPage();
-
-		await waitFor(() =>
-			expect(screen.getByText(/updated in its git repository/)).toBeInTheDocument(),
-		);
-		expect(screen.queryByText('Restart to update')).toBeNull();
-	});
-
-	it('shows no banner on a local notebook even when versions differ', async () => {
-		makeFetch({
-			role: 'editor',
-			session: gitEditSession({ source_version_id: 'ver-old' }),
-			headVersion: 'ver-head',
-		});
-		const { container } = renderPage();
-
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-		expect(screen.queryByText(/updated in its git repository/)).toBeNull();
-	});
-
-	it('Restart to update confirms (cancel is a no-op), then tears down and starts fresh', async () => {
-		const user = userEvent.setup();
-		const impl = makeFetch({
-			role: 'editor',
-			sourceType: 'git',
-			session: gitEditSession({ source_version_id: 'ver-old' }),
-			headVersion: 'ver-head',
-		});
-		const { container } = renderPage();
-		await waitFor(() => expect(screen.getByText('Restart to update')).toBeInTheDocument());
-
-		await user.click(screen.getByText('Restart to update'));
-		// The dialog, not a teardown, is what a click produces.
-		expect(sessionPosts(impl)).toHaveLength(1);
-		let dialog = await screen.findByRole('dialog');
-		expect(within(dialog).getByText(/latest synced version/)).toBeInTheDocument();
-
-		await user.click(within(dialog).getByRole('button', { name: 'Cancel' }));
-		expect(impl.mock.calls.some(([, init]) => init?.method === 'DELETE')).toBe(false);
-		expect(container.querySelector('iframe')).not.toBeNull();
-
-		await user.click(screen.getByText('Restart to update'));
-		dialog = await screen.findByRole('dialog');
-		await user.click(within(dialog).getByRole('button', { name: 'Restart' }));
-		await waitFor(() => expect(sessionPosts(impl)).toHaveLength(2));
-		const deletes = impl.mock.calls.filter(([, init]) => init?.method === 'DELETE');
-		expect(deletes).toHaveLength(1);
-	});
-
-	it('lets a manager open a pull request from a persistent running session', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-		});
-		const popup = {
-			opener: window,
-			location: { href: 'about:blank' },
-			close: vi.fn(),
-		};
-		vi.spyOn(window, 'open').mockReturnValue(popup as unknown as Window);
-		renderPage();
-
-		await user.click(await screen.findByRole('button', { name: 'Open PR' }));
-		await waitFor(() => expect(popup.location.href).toBe('https://github.com/org/repo/pull/17'));
-		expect(popup.opener).toBeNull();
-		expect(
-			fetch.mock.calls.some(
-				([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST',
-			),
-		).toBe(true);
-		expect(await screen.findByRole('button', { name: 'View PR' })).toBeEnabled();
-	});
-
-	it('views the published PR without creating another proposal', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-		});
-		const pendingPopup = {
-			opener: window,
-			location: { href: 'about:blank' },
-			close: vi.fn(),
-		};
-		const viewPopup = { opener: window };
-		vi.spyOn(window, 'open')
-			.mockReturnValueOnce(pendingPopup as unknown as Window)
-			.mockReturnValueOnce(viewPopup as unknown as Window);
-		renderPage();
-
-		await user.click(await screen.findByRole('button', { name: 'Open PR' }));
-		const viewButton = await screen.findByRole('button', { name: 'View PR' });
-		await user.click(viewButton);
-
-		expect(window.open).toHaveBeenLastCalledWith('https://github.com/org/repo/pull/17', '_blank');
-		expect(viewPopup.opener).toBeNull();
-		expect(
-			fetch.mock.calls.filter(
-				([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST',
-			),
-		).toHaveLength(1);
-	});
-
-	it('updates the current PR and replaces it only when creating a new PR', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-		});
-		const firstPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const updatePopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const newPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const viewPopup = { opener: window };
-		vi.spyOn(window, 'open')
-			.mockReturnValueOnce(firstPopup as unknown as Window)
-			.mockReturnValueOnce(updatePopup as unknown as Window)
-			.mockReturnValueOnce(newPopup as unknown as Window)
-			.mockReturnValueOnce(viewPopup as unknown as Window);
-		renderPage();
-
-		await user.click(await screen.findByRole('button', { name: 'Open PR' }));
-		await screen.findByRole('button', { name: 'View PR' });
-		await user.click(screen.getByRole('button', { name: 'pull request options' }));
-		await user.click(await screen.findByRole('menuitem', { name: 'Update PR' }));
-		await waitFor(() =>
-			expect(updatePopup.location.href).toBe('https://github.com/org/repo/pull/17'),
-		);
-
-		await user.click(screen.getByRole('button', { name: 'pull request options' }));
-		await user.click(await screen.findByRole('menuitem', { name: 'Create new PR' }));
-		await waitFor(() => expect(newPopup.location.href).toBe('https://github.com/org/repo/pull/18'));
-		await user.click(screen.getByRole('button', { name: 'View PR' }));
-		expect(window.open).toHaveBeenLastCalledWith('https://github.com/org/repo/pull/18', '_blank');
-
-		const requests = fetch.mock.calls.filter(
-			([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST',
-		);
-		expect(requests).toHaveLength(3);
-		expect(JSON.parse(String(requests[0]?.[1]?.body))).not.toHaveProperty('target_proposal_id');
-		expect(JSON.parse(String(requests[1]?.[1]?.body))).toMatchObject({
-			target_proposal_id: 'prop-1234567890abcdef',
-		});
-		expect(JSON.parse(String(requests[2]?.[1]?.body))).not.toHaveProperty('target_proposal_id');
-		const keys = requests.map(([, init]) => new Headers(init?.headers).get('idempotency-key'));
-		expect(new Set(keys).size).toBe(3);
-	});
-
-	it('targets each subsequent update at the latest published proposal', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-		});
-		vi.spyOn(window, 'open').mockImplementation(
-			() =>
-				({
-					opener: window,
-					location: { href: 'about:blank' },
-					close: vi.fn(),
-				}) as unknown as Window,
-		);
-		renderPage();
-
-		await user.click(await screen.findByRole('button', { name: 'Open PR' }));
-		await screen.findByRole('button', { name: 'View PR' });
-		for (let index = 0; index < 2; index++) {
-			await user.click(screen.getByRole('button', { name: 'pull request options' }));
-			await user.click(await screen.findByRole('menuitem', { name: 'Update PR' }));
-			await waitFor(() =>
-				expect(
-					fetch.mock.calls.filter(
-						([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST',
-					),
-				).toHaveLength(index + 2),
-			);
-			if (index === 0) {
-				await waitFor(() =>
-					expect(screen.getByRole('button', { name: 'pull request options' })).toBeEnabled(),
-				);
-			}
-		}
-
-		const requests = fetch.mock.calls.filter(
-			([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST',
-		);
-		expect(JSON.parse(String(requests[1]?.[1]?.body))).toMatchObject({
-			target_proposal_id: 'prop-1234567890abcdef',
-		});
-		expect(JSON.parse(String(requests[2]?.[1]?.body))).toMatchObject({
-			target_proposal_id: 'prop-2234567890abcdef',
-		});
-	});
-
-	it('keeps the current PR and idempotency key when an update transiently fails', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-			changeRequestFailOn: [2],
-		});
-		const openPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const failedPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const retryPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		vi.spyOn(window, 'open')
-			.mockReturnValueOnce(openPopup as unknown as Window)
-			.mockReturnValueOnce(failedPopup as unknown as Window)
-			.mockReturnValueOnce(retryPopup as unknown as Window);
-		renderPage();
-
-		await user.click(await screen.findByRole('button', { name: 'Open PR' }));
-		await screen.findByRole('button', { name: 'View PR' });
-		await user.click(screen.getByRole('button', { name: 'pull request options' }));
-		await user.click(await screen.findByRole('menuitem', { name: 'Update PR' }));
-		await waitFor(() => expect(failedPopup.close).toHaveBeenCalledOnce());
-		expect(screen.getByRole('button', { name: 'View PR' })).toBeEnabled();
-		await user.click(screen.getByRole('button', { name: 'pull request options' }));
-		await user.click(await screen.findByRole('menuitem', { name: 'Update PR' }));
-		await waitFor(() =>
-			expect(retryPopup.location.href).toBe('https://github.com/org/repo/pull/17'),
-		);
-
-		const requests = fetch.mock.calls.filter(
-			([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST',
-		);
-		const updateRequests = requests.slice(1);
-		expect(updateRequests).toHaveLength(2);
-		const firstUpdateKey = new Headers(updateRequests[0]?.[1]?.headers).get('idempotency-key');
-		expect(firstUpdateKey).not.toBeNull();
-		expect(new Headers(updateRequests[1]?.[1]?.headers).get('idempotency-key')).toBe(
-			firstUpdateKey,
-		);
-		expect(updateRequests.map(([, init]) => changeRequestBody(init).target_proposal_id)).toEqual([
-			'prop-1234567890abcdef',
-			'prop-1234567890abcdef',
-		]);
-	});
-
-	it('keeps the old PR selected when creating a replacement PR fails', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-			changeRequestFailOn: [2],
-		});
-		const openPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const failedPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const viewPopup = { opener: window };
-		vi.spyOn(window, 'open')
-			.mockReturnValueOnce(openPopup as unknown as Window)
-			.mockReturnValueOnce(failedPopup as unknown as Window)
-			.mockReturnValueOnce(viewPopup as unknown as Window);
-		renderPage();
-
-		await user.click(await screen.findByRole('button', { name: 'Open PR' }));
-		await screen.findByRole('button', { name: 'View PR' });
-		await user.click(screen.getByRole('button', { name: 'pull request options' }));
-		await user.click(await screen.findByRole('menuitem', { name: 'Create new PR' }));
-		await waitFor(() => expect(failedPopup.close).toHaveBeenCalledOnce());
-		await user.click(screen.getByRole('button', { name: 'View PR' }));
-
-		expect(window.open).toHaveBeenLastCalledWith('https://github.com/org/repo/pull/17', '_blank');
-		expect(viewPopup.opener).toBeNull();
-		expect(
-			fetch.mock.calls.filter(
-				([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST',
-			),
-		).toHaveLength(2);
-	});
-
-	it('rotates only the failed update key when its proposal must be recaptured', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-			changeRequestFailOn: [2],
-			changeRequestFailure: {
-				code: 'PROPOSAL_RETRY_REQUIRED',
-				message: 'The proposal payload expired; retry with a new idempotency key',
-				status: 409,
-			},
-		});
-		vi.spyOn(window, 'open').mockImplementation(
-			() =>
-				({
-					opener: window,
-					location: { href: 'about:blank' },
-					close: vi.fn(),
-				}) as unknown as Window,
-		);
-		renderPage();
-
-		await user.click(await screen.findByRole('button', { name: 'Open PR' }));
-		await screen.findByRole('button', { name: 'View PR' });
-		for (let index = 0; index < 2; index++) {
-			await user.click(screen.getByRole('button', { name: 'pull request options' }));
-			await user.click(await screen.findByRole('menuitem', { name: 'Update PR' }));
-			await waitFor(() =>
-				expect(
-					fetch.mock.calls.filter(
-						([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST',
-					),
-				).toHaveLength(index + 2),
-			);
-			if (index === 0) {
-				await waitFor(() =>
-					expect(screen.getByRole('button', { name: 'pull request options' })).toBeEnabled(),
-				);
-			}
-		}
-
-		const updates = fetch.mock.calls
-			.filter(([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST')
-			.slice(1);
-		expect(updates).toHaveLength(2);
-		expect(new Headers(updates[0]?.[1]?.headers).get('idempotency-key')).not.toBe(
-			new Headers(updates[1]?.[1]?.headers).get('idempotency-key'),
-		);
-		expect(updates.map(([, init]) => changeRequestBody(init).target_proposal_id)).toEqual([
-			'prop-1234567890abcdef',
-			'prop-1234567890abcdef',
-		]);
-	});
-
-	it('navigates the current tab when the pull-request popup is blocked', async () => {
-		const user = userEvent.setup();
-		makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-		});
-		const assign = vi.fn();
-		vi.stubGlobal('location', { ...window.location, assign });
-		const open = vi.spyOn(window, 'open').mockReturnValue(null);
-		renderPage();
-
-		await user.click(await screen.findByRole('button', { name: 'Open PR' }));
-		await waitFor(() => expect(assign).toHaveBeenCalledWith('https://github.com/org/repo/pull/17'));
-		expect(open).toHaveBeenCalledOnce();
-		expect(open).toHaveBeenCalledWith('about:blank', '_blank');
-	});
-
-	it('reuses the idempotency key when a manager retries a failed publication', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-			changeRequestFailures: 1,
-		});
-		const firstPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const secondPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		vi.spyOn(window, 'open')
-			.mockReturnValueOnce(firstPopup as unknown as Window)
-			.mockReturnValueOnce(secondPopup as unknown as Window);
-		renderPage();
-
-		const button = await screen.findByRole('button', { name: 'Open PR' });
-		await user.click(button);
-		await waitFor(() => expect(firstPopup.close).toHaveBeenCalledOnce());
-		await waitFor(() => expect(button).toBeEnabled());
-		await user.click(button);
-		await waitFor(() =>
-			expect(secondPopup.location.href).toBe('https://github.com/org/repo/pull/17'),
-		);
-
-		const keys = fetch.mock.calls
-			.filter(([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST')
-			.map(([, init]) => new Headers(init?.headers).get('idempotency-key'));
-		expect(keys).toHaveLength(2);
-		expect(keys[1]).toBe(keys[0]);
-	});
-
-	it('rotates the idempotency key when the proposal must be recaptured', async () => {
-		const user = userEvent.setup();
-		const fetch = makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-			changeRequestFailures: 1,
-			changeRequestFailure: {
-				code: 'PROPOSAL_RETRY_REQUIRED',
-				message: 'The proposal payload expired; retry with a new idempotency key',
-				status: 409,
-			},
-		});
-		const firstPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		const secondPopup = { opener: window, location: { href: 'about:blank' }, close: vi.fn() };
-		vi.spyOn(window, 'open')
-			.mockReturnValueOnce(firstPopup as unknown as Window)
-			.mockReturnValueOnce(secondPopup as unknown as Window);
-		renderPage();
-
-		const button = await screen.findByRole('button', { name: 'Open PR' });
-		await user.click(button);
-		await waitFor(() => expect(firstPopup.close).toHaveBeenCalledOnce());
-		await waitFor(() => expect(button).toBeEnabled());
-		await user.click(button);
-		await waitFor(() =>
-			expect(secondPopup.location.href).toBe('https://github.com/org/repo/pull/17'),
-		);
-
-		const keys = fetch.mock.calls
-			.filter(([url, init]) => String(url).endsWith('/change-requests') && init?.method === 'POST')
-			.map(([, init]) => new Headers(init?.headers).get('idempotency-key'));
-		expect(keys).toHaveLength(2);
-		expect(keys[1]).not.toBe(keys[0]);
-	});
-
-	it('does not offer pull-request publishing to project editors', async () => {
-		makeFetch({
-			role: 'editor',
-			sourceType: 'git',
-			session: gitEditSession(),
-			sourceControlProviders: ['github'],
-		});
-		const { container } = renderPage();
-
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-		expect(screen.queryByRole('button', { name: 'Open PR' })).toBeNull();
-	});
-
-	it('handles an older capability response without source-control fields', async () => {
-		makeFetch({
-			role: 'manager',
-			sourceType: 'git',
-			session: gitEditSession(),
-			omitSourceControlCapability: true,
-		});
-		const { container } = renderPage();
-
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-		expect(screen.queryByRole('button', { name: 'Open PR' })).toBeNull();
-	});
-});
-
-describe('NotebookPage app variant', () => {
-	const appSession = (overrides: Partial<Session> = {}) =>
-		runningSession({ mode: 'app', source_version_id: 'ver-head', ...overrides });
-
-	it('starts a run session, shows app chrome, and hides edit-only affordances', async () => {
-		const impl = makeFetch({ role: 'editor', session: appSession() });
-		const { container } = renderPage('app');
-
-		await waitFor(() =>
-			expect(
-				container.querySelector('iframe[src="https://sandbox.example/kernel?theme=light"]'),
-			).not.toBeNull(),
-		);
-		const [, init] = sessionPosts(impl)[0];
-		expect(String(init?.body)).toContain('"mode":"app"');
-		expect(screen.getByText('App')).toBeInTheDocument();
-		expect(screen.getByText('Restart')).toBeInTheDocument();
-		expect(screen.getByText('Stop')).toBeInTheDocument();
-		expect(screen.queryByLabelText('Rename notebook')).toBeNull();
-	});
-
-	it('shows the staleness banner when the app trails the notebook head', async () => {
-		makeFetch({
-			role: 'editor',
-			session: appSession({ source_version_id: 'ver-old' }),
-			headVersion: 'ver-head',
-		});
-		renderPage('app');
-
-		await waitFor(() => expect(screen.getByText(/serving an older version/)).toBeInTheDocument());
-		expect(screen.getByText('Restart to update')).toBeInTheDocument();
-	});
-
-	it('suppresses the staleness banner while the notebook is being edited', async () => {
-		makeFetch({
-			role: 'editor',
-			session: appSession({ source_version_id: 'ver-old' }),
-			headVersion: 'ver-head',
-			projectSessions: [runningSession({ session_id: 'sess-edit', mode: 'edit' })],
-		});
-		const { container } = renderPage('app');
-
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-		await waitFor(() => expect(screen.queryByText(/serving an older version/)).toBeNull());
-	});
-
-	it('does not suppress the staleness banner for a temporary editor', async () => {
-		makeFetch({
-			role: 'editor',
-			session: appSession({ source_version_id: 'ver-old' }),
-			headVersion: 'ver-head',
-			projectSessions: [runningSession({ session_id: 'sess-edit', mode: 'edit', ephemeral: true })],
-		});
-		renderPage('app');
-
-		await waitFor(() => expect(screen.getByText(/serving an older version/)).toBeInTheDocument());
-	});
-
-	it('does not suppress the banner during editing on a git-synced notebook', async () => {
-		makeFetch({
-			role: 'editor',
-			sourceType: 'git',
-			session: appSession({ source_version_id: 'ver-old' }),
-			headVersion: 'ver-head',
-			projectSessions: [runningSession({ session_id: 'sess-edit', mode: 'edit' })],
-		});
-		renderPage('app');
-
-		await waitFor(() => expect(screen.getByText(/serving an older version/)).toBeInTheDocument());
-	});
-
-	it('shows no staleness banner when the app serves the head version', async () => {
-		const { container } = (() => {
-			makeFetch({ role: 'editor', session: appSession(), headVersion: 'ver-head' });
-			return renderPage('app');
-		})();
-
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-		expect(screen.queryByText(/serving an older version/)).toBeNull();
-	});
-
-	it('viewer + applications: uses the app but gets no Stop/Restart controls', async () => {
-		const impl = makeFetch({
-			role: 'viewer',
-			viewerMode: 'applications',
-			session: appSession({ can: { attach: true, stop: false } }),
-		});
-		const { container } = renderPage('app');
-
-		await waitFor(() =>
-			expect(
-				container.querySelector('iframe[src="https://sandbox.example/kernel?theme=light"]'),
-			).not.toBeNull(),
-		);
-		expect(sessionPosts(impl)).toHaveLength(1);
-		expect(screen.getByText('App')).toBeInTheDocument();
-		expect(screen.queryByText('Restart')).toBeNull();
-		expect(screen.queryByText('Stop')).toBeNull();
-	});
-
-	it('viewer + applications: the staleness banner has no restart CTA', async () => {
-		makeFetch({
-			role: 'viewer',
-			viewerMode: 'applications',
-			session: appSession({ source_version_id: 'ver-old', can: { attach: true, stop: false } }),
-			headVersion: 'ver-head',
-		});
-		renderPage('app');
-
-		await waitFor(() => expect(screen.getByText(/serving an older version/)).toBeInTheDocument());
-		// Restarting the shared app is editor-only; a viewer clicking through
-		// would 403 on the stop half.
-		expect(screen.queryByText('Restart to update')).toBeNull();
-	});
-
-	it('Restart confirms before disconnecting everyone (with the connection hint)', async () => {
-		const user = userEvent.setup();
-		const impl = makeFetch({
-			role: 'editor',
-			session: appSession({ active_connections: 3 }),
-		});
-		const { container } = renderPage('app');
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-
-		await user.click(screen.getByText('Restart'));
-		// The dialog, not a teardown, is what a click produces.
-		expect(impl.mock.calls.some(([, init]) => init?.method === 'DELETE')).toBe(false);
-		const dialog = await screen.findByRole('dialog');
-		expect(within(dialog).getByText(/About 3 people are connected/)).toBeInTheDocument();
-
-		await user.click(within(dialog).getByRole('button', { name: 'Restart' }));
-		await waitFor(() =>
-			expect(impl.mock.calls.some(([, init]) => init?.method === 'DELETE')).toBe(true),
-		);
-	});
-
-	it('Stop confirms too; cancel leaves the app untouched', async () => {
-		const user = userEvent.setup();
-		const impl = makeFetch({ role: 'editor', session: appSession() });
-		const { container } = renderPage('app');
-		await waitFor(() => expect(container.querySelector('iframe')).not.toBeNull());
-
-		await user.click(screen.getByText('Stop'));
-		const dialog = await screen.findByRole('dialog');
-		expect(within(dialog).getByText(/Anyone using it will be disconnected/)).toBeInTheDocument();
-
-		await user.click(within(dialog).getByRole('button', { name: 'Cancel' }));
-		expect(impl.mock.calls.some(([, init]) => init?.method === 'DELETE')).toBe(false);
-		expect(container.querySelector('iframe')).not.toBeNull();
-	});
-
-	// `sandbox_url` is withheld from a caller the kernel gates would reject.
-	it('a running app the caller cannot reach renders the access-ended panel', async () => {
-		makeFetch({ role: 'viewer', session: appSession({ sandbox_url: undefined }) });
-		const { container } = renderPage('app');
-
-		await waitFor(() => expect(screen.getByText('Access ended')).toBeInTheDocument());
-		expect(screen.getByText(/no longer have access/)).toBeInTheDocument();
-		expect(container.querySelector('iframe')).toBeNull();
-		// The app is still serving everyone else, so a restart could only 403.
-		expect(screen.queryByText('Restart app')).toBeNull();
-		expect(screen.getByText('Back')).toBeInTheDocument();
-	});
-
-	it('a viewer’s 403 renders the error panel without a Retry button', async () => {
-		const impl = makeFetch({
-			role: 'viewer',
-			computeProfile: 'large',
-			computeProfileOverride: 'editors',
-			computeProfiles: [
-				{ name: 'small', cpu: 1 },
-				{ name: 'large', cpu: 8 },
-			],
-			createError: { code: 'FORBIDDEN', message: "Requires 'editor' role", status: 403 },
-		});
-		renderPage('app');
-
-		await waitFor(() => expect(screen.getByText(/Requires 'editor' role/)).toBeInTheDocument());
-		expect(screen.queryByText('Retry')).toBeNull();
-		expect(screen.queryByText('Retry with Default')).toBeNull();
-		expect(screen.getByText('Back')).toBeInTheDocument();
-		// The doomed request fired once — no loop.
-		expect(sessionPosts(impl)).toHaveLength(1);
-	});
-
-	it('does not offer a Default bypass for a failed shared app', async () => {
-		makeFetch({
-			role: 'viewer',
-			viewerMode: 'applications',
-			computeProfile: 'large',
-			computeProfileOverride: 'editors',
-			computeProfiles: [
-				{ name: 'small', cpu: 1 },
-				{ name: 'large', cpu: 8 },
-			],
-			createError: {
-				code: 'RESOURCE_EXHAUSTED',
-				message: 'No nodes can schedule this profile',
-				status: 429,
-			},
-		});
-		renderPage('app');
-
-		expect(await screen.findByText('No nodes can schedule this profile')).toBeInTheDocument();
-		expect(screen.queryByText('Retry with Default')).toBeNull();
-		expect(screen.getByRole('button', { name: 'Retry' })).toBeInTheDocument();
-	});
-
-	it('offers a one-shot Retry with Default without replacing the stored profile', async () => {
-		const user = userEvent.setup();
-		const impl = makeFetch({
-			role: 'editor',
-			computeProfile: 'large',
-			computeProfileOverride: 'editors',
-			computeProfiles: [
-				{ name: 'small', cpu: 1 },
-				{ name: 'large', cpu: 8 },
-			],
-			createError: {
-				code: 'RESOURCE_EXHAUSTED',
-				message: 'No nodes can schedule this profile',
-				status: 429,
-			},
-		});
-		renderPage();
-
-		expect(await screen.findByText('No nodes can schedule this profile')).toBeInTheDocument();
-		expect(screen.queryByText(/a larger profile may be needed/)).not.toBeInTheDocument();
-		await user.click(screen.getByRole('button', { name: 'Retry with Default' }));
-
-		await waitFor(() => expect(sessionPosts(impl)).toHaveLength(2));
-		const retryBody = sessionPosts(impl)[1][1]?.body;
-		expect(JSON.parse(String(retryBody))).toEqual({ compute_profile: 'default' });
-		expect(screen.queryByRole('button', { name: 'Retry with Default' })).toBeNull();
 	});
 });
